move simpleConfigManager aside
